[ style ] 이메일 코드 컴포넌트 분리
